Add geocoder repository and update config for Flyway and jOOQ
This commit adds the geocoder repository and updates the project configuration to integrate with Flyway and jOOQ. This includes generating relevant jOOQ classes and adding a build.gradle.kts configuration file for the repository. The configuration update enables Flyway database migrations and generates sources with jOOQ. Additionally added an implementation for `GeocoderSolutionJooqAdapter` and modified/included several other Kotlin files across the codebase.