Remove draggable functionality from map markers
The code changes remove the functions and events associated with dragging the markers on the map. The draggable property of the markers was set to false, the 'movePoint' event was removed, and the functions handling the marker dragging effects were also removed. These changes imply that the markers are now static and users will no longer be able to move them.